refactor: Restructure core module into submodules for better organization
