+++ src/core/gui/mod.rs
@@ -1,0 +1,1 @@
+pub mod gui_manager;
