+++ src/core/infrastructure/io_manager.rs
@@ -1,4 +1,4 @@
-use crate::core::app_config::AppConfig;
+use crate::core::infrastructure::app_config::AppConfig;
 use crate::interface::file_system::FileSystemTrait;
 use crate::platform::file_system::FileSystem;
 use std::ops::Deref;
