+++ src/core/infrastructure/mod.rs
@@ -1,0 +1,3 @@
+pub mod app_config;
+pub mod database_manager;
+pub mod io_manager;
