+++ src/core/mod.rs
@@ -1,9 +1,6 @@
-pub mod app_config;
-pub mod backup_engine;
-pub mod database_manager;
 pub mod event_bus;
-pub mod gui_manager;
-pub mod io_manager;
-pub mod progress_tracker;
-pub mod schedule_manager;
 pub mod system;
+pub mod backup;
+pub mod schedule;
+pub mod infrastructure;
+pub mod gui;
