+++ src/core/schedule/schedule_manager.rs
@@ -1,5 +1,5 @@
-use crate::core::app_config::AppConfig;
-use crate::core::database_manager::DatabaseManager;
+use crate::core::infrastructure::app_config::AppConfig;
+use crate::core::infrastructure::database_manager::DatabaseManager;
 use crate::core::event_bus::EventBus;
 use crate::interface::repository::schedule::ScheduleRepository;
 use crate::interface::service_unit::ServiceUnit;
