+++ src/core/system.rs
@@ -1,11 +1,11 @@
-use crate::core::app_config::AppConfig;
-use crate::core::backup_engine::BackupEngine;
-use crate::core::database_manager::DatabaseManager;
+use crate::core::infrastructure::app_config::AppConfig;
+use crate::core::backup::backup_engine::BackupEngine;
+use crate::core::infrastructure::database_manager::DatabaseManager;
 use crate::core::event_bus::EventBus;
-use crate::core::gui_manager::GuiManager;
-use crate::core::io_manager::IOManager;
-use crate::core::progress_tracker::ProgressTracker;
-use crate::core::schedule_manager::ScheduleManager;
+use crate::core::gui::gui_manager::GuiManager;
+use crate::core::infrastructure::io_manager::IOManager;
+use crate::core::backup::progress_tracker::ProgressTracker;
+use crate::core::schedule::schedule_manager::ScheduleManager;
 use crate::interface::service_unit::ServiceUnit;
 use crate::model::error::Error;
 use crate::model::error::system::SystemError;
